% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_parameters_table.R
\name{compile_parameters_table}
\alias{compile_parameters_table}
\title{Script to apply the reflectance_color_indices function to a folder}
\usage{
compile_parameters_table(my_folder)
}
\arguments{
\item{my_folder}{path to folder excel files}
}
\value{
a list with lightness, a*, b*, chroma and hue table (tibble). It also produces an excell file with all the values LAB_LCH_parameters_full_table.xlsx.
}
\description{
Script to compile all the excel files produced by the script process_lab_lch_folder. It uses the name of the files to extract information about the replicate number and the sampling day. Only applicable to Konstantina's workflow.
}
\keyword{external}
