% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stats.R
\name{extract_stats}
\alias{extract_stats}
\title{Calculate several statistics from selected ROI}
\usage{
extract_stats(tif_file, roi_file, type_file)
}
\arguments{
\item{tif_file}{path to tif file with the parameters to extract}

\item{roi_file}{path shapefile with the ROI file}

\item{type_file}{parameter to setup if the file is LAB or LCH.}
}
\value{
The function returns a list with the statistics and the pixel values for each ROI.
}
\description{
This function calculates statistics from six ROI stored in a shapefile. It was designed for Konstantina's workflow and it is unlikely to work in another context. The ROI need to be called D1, D2, V1, V2, eye, gills.
}
\keyword{external}
