% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labfunctions.R
\name{lab_function}
\alias{lab_function}
\title{Construct L*a*b* images from a reflectance spectra}
\usage{
lab_function(wl = NA, spectra = NA, file = NA)
}
\arguments{
\item{wl}{vector with the wavelengths}

\item{spectra}{vector with the reflectance data}

\item{file}{string with the path for an ASD reflectance file}
}
\value{
The function returns a list with: 1) a dataframe with the input data (wl and spectra); 2) a vector with the XYZ parameters; and 3) a vector with the L*a*b* parameters
}
\description{
This function takes a reflectance spectra (R object ot file) as an input and calculates the corresponding L*a*b* parameters. TO DO - add info about the models and illuminants used to estimate XYZ and L*a*b*.
}
\keyword{external}
