% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterfunctions.R
\name{lab_raster}
\alias{lab_raster}
\title{Construct L*a*b* images from an hyperspectral camera}
\usage{
lab_raster(raster)
}
\arguments{
\item{raster}{Raster object with the camera file}
}
\value{
The function returns a raster stack object with three layers: L*,a* and b*
}
\description{
This function takes as input an hyperspectral image and calculates the corresponding L*a*b* parameters for each pixel, result in a stack of three images corresponding the each of the  L*a*b* parameters
}
\keyword{external}
