% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lch_raster.R
\name{lch_raster}
\alias{lch_raster}
\title{Construct Lch images from an hyperspectral camera}
\usage{
lch_raster(raster)
}
\arguments{
\item{raster}{Raster object with the camera file}
}
\value{
The function returns a raster stack object with three layers: L, c and h
}
\description{
This function takes as input an hyperspectral image and calculates the corresponding Lch parameters for each pixel, result in a stack of three images corresponding the each of the  Lch parameters.
}
\keyword{external}
