% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_lab_lch_folder.R
\name{process_lab_lch_folder}
\alias{process_lab_lch_folder}
\title{Script to apply the reflectance_color_indices function to a folder}
\usage{
process_lab_lch_folder(my_folder, my_reference)
}
\arguments{
\item{my_folder}{path to folder with the cube camera files, recursive search.}

\item{my_reference}{path to the reference file}
}
\value{
Doesn't return any R object directly, only through the reflectance_color_indices function.
}
\description{
Script to apply the reflectance_color_indices function to a folder. Only applicable to Konstantina's workflow.
}
\keyword{external}
