% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectance_color_indices.R
\name{reflectance_color_indices}
\alias{reflectance_color_indices}
\title{Calculate reflectance and LAB, LCH parameters using an 18 percent reference image}
\usage{
reflectance_color_indices(
  camera_file,
  reference_file,
  roi_file,
  my_plots = TRUE
)
}
\arguments{
\item{camera_file}{path to the camera file, cube version}

\item{reference_file}{path to the 18 percent reference file}

\item{roi_file}{path shapefile with the ROI file}

\item{my_plots}{logical value to activate-desactivate plots. Not implemented yet}
}
\value{
The function outputs a reflectance tif file, a LAB tif file, a LCH tif file, an excel LAB file stats + pixel values, an excel LCH file stats + pixel values and it returns a R list with lab/lch statistics, the reflectance image, and the lab+lch images
}
\description{
This function calculates reflectance using a 18 percent reference image, calculates the corresponding Lab and LCH parameters for each pixel and then it extracts the pixel data using a shapefile with the desired ROI. It was designed specifically for Konstantina's workflow so it is unlikelly to function outside that scope.
}
\keyword{external}
